Check maximum number of characters before urlencode
We need to check the maximum length of the character before encoding the text as the urlencode changes the number of character for special characters. 

For example `foo & bar` will change to `foo%20%26%20bar%20` and as a result the number of characters would increase. The API accepts the number of characters when encoded.